Move CustomTransformer plugin test to specific test file
